feat: restructure database crate
