+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
     "crates/interpreter",
     "crates/precompile",
     "crates/optimism",
+    "crates/database",
     "crates/database/interface",
     "crates/bytecode",
     "crates/state",
@@ -19,6 +20,7 @@ default-members = ["crates/revm"]
 # revm
 primitives = { path = "crates/primitives", package = "revm-primitives", version = "9.0.1", default-features = false }
 bytecode = { path = "crates/bytecode", package = "revm-bytecode", version = "1.0.0", default-features = false }
+database = { path = "crates/database", package = "revm-database", version = "1.0.0", default-features = false }
 database-interface = { path = "crates/database/interface", package = "revm-database-interface", version = "1.0.0", default-features = false }
 specification = { path = "crates/specification", package = "revm-specification", version = "1.0.0", default-features = false }
 state = { path = "crates/state", package = "revm-state", version = "1.0.0", default-features = false }
