+++ bins/revme/src/cmd/statetest/merkle_trie.rs
@@ -1,10 +1,8 @@
 use alloy_rlp::{RlpEncodable, RlpMaxEncodedLen};
+use database::PlainAccount;
 use hash_db::Hasher;
 use plain_hasher::PlainHasher;
-use revm::{
-    db::PlainAccount,
-    primitives::{keccak256, Address, Log, B256, U256},
-};
+use revm::primitives::{keccak256, Address, Log, B256, U256};
 use triehash::sec_trie_root;
 
 pub fn log_rlp_hash(logs: &[Log]) -> B256 {
