+++ crates/optimism/Cargo.toml
@@ -37,6 +37,7 @@ serde = { version = "1.0", default-features = false, features = [
 
 
 [dev-dependencies]
+database.workspace = true
 anyhow = "1.0.89"
 criterion = "0.5"
 indicatif = "0.17"
